+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - ControlSystem - only add switcher control if control system has inputs/outputs
  - ControlSystemSwitcherControl - update video source detection on input port only for video detection changed
  - ControlSystemSwitcherControl - Breakaway routing check feedback bools for supported instead of set bools
+ - SwitcherTelemetry - Removed methods for Output Mute/Unmute/Volume that weren't implemented
+
+### Removed
+ - SwitcherTelemetry - AudioBreakawayEnabled and UsbBreakawayEnabled remvoed
 
 ## [13.13.2] - 2020-03-05
 ### Changed